+++ psl-core/src/main/java/org/linqs/psl/application/learning/weight/search/grid/InitialWeightRandomGridSearch.java
@@ -56,7 +56,7 @@ public class InitialWeightRandomGridSearch extends RandomGridSearch {
     }
 
     @Override
-    protected double inspectLocation(double[] weights) {
+    protected double inspectLocation(float[] weights) {
         // Just have the internal WLA learn and then get the loss as the score.
         internalWLA.learn();
