+++ psl-core/src/main/java/org/linqs/psl/model/rule/UnweightedRule.java
@@ -24,5 +24,5 @@ import org.linqs.psl.model.atom.GroundAtom;
  * which constrain the values that {@link GroundAtom GroundAtoms} can take.
  */
 public interface UnweightedRule extends Rule {
-    public WeightedRule relax(double weight, boolean squared);
+    public WeightedRule relax(float weight, boolean squared);
 }
